refactor(typage): les loaders cessent de re-typer — suppression des re-CAST (#293)
ADR-0035 : le loader ne connaît plus aucun type, il trust ce que dbt émet.

- sql.py : tous les `col_cast_double` retirés (fonction supprimée). F15
  (prix/quantite/montant) et puissance_souscrite_kva étaient déjà DOUBLE côté dbt
  → no-op. Les index C15 avant/après et R64 (bigint) cessent d'être castés en DOUBLE.
- Historique : index avant/après Float64 → Int64 (kWh entiers, hérité du bigint dbt).
- Pas de transform Wh→kWh mort à retirer (déjà descendu en dbt à ADR-0034).

Le test de parité (#291) reste le garde-fou de frontière. Empreinte /releves stable
(le mart est inchangé). Fixtures aval + snapshot de structure alignés sur Int64.

+++ tests/integration/test_pipelines_snapshot.py
@@ -67,8 +67,8 @@ def historique_brut(evenements: list[dict]) -> pl.LazyFrame:
     for cadran in CADRANS:
         data[f"avant_{col_index(cadran)}"] = [None] * n
         data[f"apres_{col_index(cadran)}"] = [None] * n
-        schema_overrides[f"avant_{col_index(cadran)}"] = pl.Float64
-        schema_overrides[f"apres_{col_index(cadran)}"] = pl.Float64
+        schema_overrides[f"avant_{col_index(cadran)}"] = pl.Int64
+        schema_overrides[f"apres_{col_index(cadran)}"] = pl.Int64
 
     return pl.LazyFrame(data, schema_overrides=schema_overrides).with_columns(
         pl.col("date_evenement").dt.replace_time_zone("Europe/Paris")
